+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
+## 0.3.0 (???)
+
+### Improvements
+
+- Add functions `difference_all`, `difference_all_tiled` and `empty` (#42)
+
+### Bug fixes
+
+### Deprecations and compatibility notes
+
+- Parameter `primitivetype` of `collection_extract()` is deprecated and replaced by
+  `keep_geom_type` (#42)
+
 ## 0.2.0 (2023-07-27)
 
 ### Improvements